Add alternative read with pressure measurement and different filtering
This seems to work much better at identifying when the screen is touched.
It still needs some work, and I think for the sake of flexibility it should
support adjustable thresholding and maybe even oversampling (although on
my display there does not seem to be a lot of advantage to oversampling).